intent recognition async requests
